pyproject.toml
[tool.poetry]
name = "crsf_parser"
version = "0.1.0"
description = ""
authors = ["Alessio Morale <alessiomorale@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
construct= "^2.10"
construct-typing = "^0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyserial = "^3.5"
black = "^21.9b0"
mypy = "^0.910"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
